+++ settings.yaml.example
@@ -1,5 +1,5 @@
 host: 127.0.0.1
 port: 8000
 database_url: "postgres://db_user:db_password@localhost:5432/cjms"
-environment: dev
+environment: local
 gcp_project: proj
